Correctly write the CRD yaml
The csv-generator writes an empty CRD yaml. The reason is that the
`resources.MigrationControllerCRDs` mapp, does not contain the
"migcontroller" key.

In addition, the way to produce the CRD is a bit too complex: it reads a
CRD string, converts it to a golang object, and then marshals it back to
yaml string (while the original string is already in the required
format).

This commit fixes the issue by embedding the right CRD file, and writes
it as is, to the standard output. This is also a bit simpler solution,
that does not involve any kind of serialization.

Signed-off-by: Nahshon Unna Tsameret <nunnatsa@redhat.com>